Fix trap handling
Logic was arse backwards
Also some small performance tweaks, overall performance is still > 100us/instr

+++ cpu.h
@@ -5,14 +5,15 @@ extern jmp_buf trapbuf;
 
 namespace pdp11 {
 struct intr {
-  uint16_t vec;
-  uint16_t pri;
+  uint8_t vec;
+  uint8_t pri;
 };
 };
 
 #define ITABN 8
 
 extern pdp11::intr itab[ITABN];
+
 
 enum {
   FLAGN = 8,
@@ -38,8 +39,8 @@ void reset(void);
 void switchmode(bool newm);
 
 void trapat(uint16_t vec);
-void interrupt(uint16_t vec, uint16_t pri);
-void handleinterrupt(uint16_t vec);
+void interrupt(uint8_t vec, uint8_t pri);
+void handleinterrupt(uint8_t vec);
 
 };
